requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
flask-cors==4.0.1
numpy==1.26.4
nltk==3.8.1
keras==3.3.3
tensorflow==2.16.1
waitress==3.0.0
pymysql==1.1.0